Add basic prometheus metrics for backup status.
This requires that `prometheus-node-exporter` is installed.

If `reportprom` is set, then during the reporting phase, metrics will be written
to `${prom_textfile_dir}/backupninja.prom`, which will then be available for
scraping by prometheus.

The metrics simply report the hostname and the number of actions run, the number
of warnings, errors, fails, and halts that were produced in the backup run.

These are written to the prometheus-node-exporter directory in an atomic way.