Allow for disabling reports by email.
This diff simply adds one line:

`if [ ! -z "$reportemail" ]; then` and then it indents the if block and adds a
final `fi`.

This check used to be there, without it, its not possible to disable report
mails, which might be something desirable if you are getting alerts through the
prometheus method.